Any matching trans delta is sufficient for constrained policies
For a constrained policy, it is sufficient if any transition delta to
some successor matches one effect descriptor of the policy rule. This is
the case because the solver forces all good transitions to end in a safe
state. Hence, there is always a path to the goal for any of the matching
successors.